requirement.txt
absl-py==2.1.0
accelerate==0.32.1
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==1.3.0
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
asyncio-atexit==1.0.1
attributedict==0.3.0
attrs==23.2.0
av==12.3.0
backcall==0.2.0
backoff==2.2.1
batchgenerators==0.25
bcj-cffi==0.5.1
beautifulsoup4==4.12.3
bitsandbytes==0.39.0
black==24.8.0
bleach==6.1.0
blessings==1.7
boto3==1.34.144
botocore==1.34.144
braceexpand==0.1.7
bypy==1.8.5
cachetools==5.5.0
cattrs==24.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chumpy==0.70
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.2
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colour-runner==0.1.1
conda-pack==0.8.0
connected-components-3d==3.22.0
contextlib2==21.6.0
contourpy==1.1.1
correlation_cuda==0.0.0
coverage==7.6.1
cpm-kernels==1.0.11
crcmod==1.7
cryptography==43.0.0
cycler==0.12.1
datasets==2.20.0
decorator==5.1.1
deepdiff==8.1.1
deepspeed==0.14.4
defusedxml==0.7.1
Deprecated==1.2.14
detectron2==0.6
dill==0.3.8
distlib==0.3.9
dnspython==2.6.1
docker-pycreds==0.4.0
dominate==2.9.1
easydict==1.13
edt==2.4.1
efficientnet_pytorch==0.7.1
einops==0.8.0
email_validator==2.2.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fast-slic==0.4.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastremap==1.15.0
feather-format==0.4.1
ffmpy==0.3.2
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
fvcore==0.1.5.post20221221
gdcm==1.1
gdown==5.2.0
GeodisTK==0.1.7
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.63.2
grad-cam==1.5.4
gradio==4.38.1
gradio_client==1.1.0
grpcio==1.48.2
h11==0.14.0
h5py==3.11.0
highdicom==0.22.0
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.5
humanfriendly==10.0
hydra-core==1.3.2
idna==3.7
imageio==2.34.1
imops==0.9.1
importlib-metadata==6.11.0
importlib_resources==6.4.0
inspecta==0.1.3
iopath==0.1.9
ipykernel==5.5.5
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
jedi==0.19.1
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.3
jupyter_core==4.12.0
jupyterlab_widgets==3.0.10
keras==2.13.1
kiwisolver==1.4.5
latex2mathml==3.77.0
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.11.2
linecache2==1.0.0
lit==18.1.4
llvmlite==0.41.1
loky==3.0.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdtex2html==1.3.0
mdurl==0.1.2
MedPy==0.5.1
mistune==0.8.4
ml_collections==0.1.1
mmaction==0.5.0
mmcv==2.2.0
mmcv-full==1.7.2
mmengine==0.10.4
modelscope==1.16.1
monai==1.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
MultiScaleDeformableAttention==1.0
multivolumefile==0.2.3
mypy-extensions==1.0.0
nbconvert==5.6.1
nbformat==5.1.3
nest_asyncio==1.6.0
networkx==3.1
neurokit2==0.2.10
nibabel==5.2.1
ninja==1.11.1.1
notebook==5.7.11
numba==0.58.1
numpy==1.23.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.555.43
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.9.0.80
openpyxl==3.1.5
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
openxlab==0.1.1
opt_einsum==3.4.0
orderly-set==5.2.3
orjson==3.10.6
oss2==2.17.0
packaging==24.2
pandas==1.5.3
pandocfilters==1.5.0
parallel-experts==0.0.0
parso==0.8.4
path==16.14.0
pathspec==0.12.1
peft==0.3.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pillow-jpls==1.3.2
pip==24.2
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==5.22.0
pluggy==1.5.0
polarTransform==2.0.0
portalocker==2.10.1
ppmd-cffi==0.4.2
prometheus_client==0.20.0
prompt-toolkit==3.0.42
proplot==0.9.7
protobuf==3.20.3
psbody-mesh==0.4
psutil==5.9.8
ptflops==0.7.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py7zr==0.15.1
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pycocotools==2.0.7
pycox==0.3.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydicom==2.4.4
pydub==0.25.1
Pygments==2.18.0
pylibjpeg==2.0.1
pylibjpeg-libjpeg==2.2.0
pylibjpeg-openjpeg==2.3.0
PYMIC==0.2.4
PyOpenGL==3.1.7
pyparsing==3.1.2
pypinyin==0.53.0
pyproject-api==1.8.0
PySocks==1.7.1
python-dateutil==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==1.9.4
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.2.0
pyzstd==0.14.4
qtconsole==5.5.2
QtPy==2.4.1
qudida==0.0.4
qwen-vl-utils==0.0.8
referencing==0.35.1
regex==2024.5.15
requests==2.28.2
requests-cache==1.2.1
requests-toolbelt==1.0.0
rich==13.4.2
rootpath==0.1.1
rpds-py==0.10.6
ruff==0.5.2
s3transfer==0.10.2
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.2
scikit-video==1.1.11
scipy==1.10.1
seaborn==0.11.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.17.0
setproctitle==1.3.3
setuptools==75.3.0
shellingham==1.5.4
SimpleITK==2.3.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.2
starlette==0.37.2
statannotations==0.6.0
SwissArmyTransformer==0.4.4
sympy==1.12
synapseclient==4.3.0
tabulate==0.9.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.0
tensorboardX==2.6.2.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
terminado==0.18.1
testpath==0.6.0
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.4.12
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.0.1
torch-geometric==2.6.1
torch-scatter==2.1.2
torch-sparse==0.6.18
torchaudio==2.0.2
torchmetrics==1.4.0.post0
torchtuples==0.2.2
torchvision==0.15.2
tornado==6.1
tox==4.23.2
tqdm==4.65.2
traceback2==1.4.0
traitlets==5.0.0
transformers==4.45.0.dev0
triton==2.0.0
ttach==0.0.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
unittest2==1.1.0
url-normalize==1.4.3
urllib3==1.26.19
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.28.0
visdom==0.2.4
vtk==9.4.0
wandb==0.18.5
watchfiles==0.22.0
wcwidth==0.2.13
webdataset==0.2.86
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==2.1.2
wheel==0.43.0
widgetsnbextension==4.0.10
wrapt==1.16.0
xpinyin==0.7.6
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
